requirements.txt
pyserial
numpy
matplotlib
requests
PySimpleGUI